+++ packages/client/src/Client.ts
@@ -74,11 +74,23 @@ class BaseEVMLC extends AbstractClient {
 	}
 
 	public async getPOAContract(): Promise<IPOAContract> {
-		return JSONBig.parse(await this.get(`/poa`));
+		const res = await this.get(`/poa`);
+
+		try {
+			return JSONBig.parse(res);
+		} catch (e) {
+			return Promise.reject(res);
+		}
 	}
 
 	public async getReceipt(txHash: string): Promise<IReceipt> {
-		return JSONBig.parse(await this.get(`/tx/${txHash}`));
+		const res = await this.get(`/tx/${txHash}`);
+
+		try {
+			return JSONBig.parse(res);
+		} catch (e) {
+			return Promise.reject(res);
+		}
 	}
 
 	public async getAccount(address: string): Promise<IBaseAccount> {
@@ -89,17 +101,35 @@ class BaseEVMLC extends AbstractClient {
 	}
 
 	public async getInfo(): Promise<any> {
-		return JSONBig.parse(await this.get('/info'));
+		const res = await this.get('/info');
+
+		try {
+			return JSONBig.parse(res);
+		} catch (e) {
+			return Promise.reject(res);
+		}
 	}
 
 	// call tx
 	public async callTx(tx: string): Promise<ICallTxResponse> {
-		return JSONBig.parse(await this.post('/call', tx));
+		const res = await this.post('/call', tx);
+
+		try {
+			return JSONBig.parse(res);
+		} catch (e) {
+			return Promise.reject(res);
+		}
 	}
 
 	// send tx
 	public async sendTx(signedTx: string): Promise<IReceipt> {
-		return JSONBig.parse(await this.post('/rawtx', signedTx));
+		const res = await this.post('/rawtx', signedTx);
+
+		try {
+			return JSONBig.parse(res);
+		} catch (e) {
+			return Promise.reject(res);
+		}
 	}
 }